add ProdcutService

+++ server/src/main/java/com/example/server/DTO/request/ProductRequest.java
@@ -18,10 +18,10 @@ public class ProductRequest {
     private String name;
     
     @NotNull @Min(0)
-    private Long sellingPrice;
+    private Integer sellingPrice;
     
     @NotNull @Min(0)
-    private Long originalPrice;
+    private Integer originalPrice;
     
     private String image;
     private String description;
